backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiomysql==0.2.0
motor==3.4.0
pymongo==4.7.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-socketio==5.11.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.0
scikit-learn==1.4.0
numpy==1.26.3
joblib==1.3.2
python-multipart==0.0.6
aiosmtplib==3.0.1
